requirements.txt
pandas
requests
streamlit
pillow
